Don't perform the same action two times + n option for 'no dependancy check'

+++ functions/67952747-d79b-45fd-8678-62be0bebb822.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
 GUID='67952747-d79b-45fd-8678-62be0bebb822'
-DESCRIPTION='RaspberryPi GPIO Python Package (Raspbian Stretch Lite)'
+DESCRIPTION='RaspberryPi GPIO Python Package'
 DEPENDS_ON=( )
-sudo apt-get install python-pip
-pip install RPi.GPIO
-sudo apt-get install picap
-picap-setup
+
+[ `dpkg -l | grep -E '^ii' | grep python-pip | wc -l` -eq 0 ] && sudo apt-get install python-pip
+[ `pip show RPi.GPIO | wc -l` -eq 0 ] && pip install RPi.GPIO
+[ `dpkg -l | grep -E '^ii' | grep picap | wc -l` -eq 0 ] && sudo apt-get install picap && picap-setup
